minor fixes to examples, updated nanobind version

+++ src/python/render_test_2.py
@@ -137,7 +137,7 @@ class MyScreen(Screen):
             )
 
             mvp = proj @ view @ model
-            self.shader.set_buffer("mvp", np.float32(mvp).T)
+            self.shader.set_buffer("mvp", mvp.T)
             with self.shader:
                 self.shader.draw_array(Shader.PrimitiveType.Triangle,
                                        0, 36, indexed=True)
